refactor: split up TestRunsPage.module.css
