+++ src/ui/components/Library/Team/View/TestRuns/Overview/TestResultListItem.module.css
@@ -1,0 +1,60 @@
+:root:global(.theme-dark) .recordingLink:hover {
+  background-color: #1c2534;
+}
+
+:root:global(.theme-light) .recordingLink:hover {
+  background-color: var(--grey-10);
+}
+
+.recordingLink {
+  display: flex;
+  width: 100%;
+  flex-grow: 1;
+  cursor: pointer;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+  gap: 0.5rem;
+  overflow: hidden;
+  border-radius: 0.375rem;
+  padding: 4px;
+}
+
+.fileInfo {
+  display: flex;
+  flex-shrink: 1;
+  flex-grow: 1;
+  flex-direction: column;
+  overflow: hidden;
+}
+
+.title {
+  display: block;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+}
+
+.filePath {
+  display: block;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  color: var(--color-dimmer);
+  font-size: 0.75rem;
+}
+
+.comments {
+  display: flex;
+  align-items: center;
+  flex-shrink: 0;
+  flex-direction: row;
+  gap: 1rem;
+  color: var(--body-color);
+}
+
+.commentIcon {
+  width: 1rem;
+}
+
+
